feat: Add navigation events on mouse-3 and esc, feat: Add custom offsets to FancyIcon
chore: Start work on actually splitting up config into separate categories
