+++ src/renderer/components/FancyIcon.vue
@@ -1,17 +1,17 @@
 <template>
     <div class="flex items-center justify-center relative h-full">
-        <Icon class="size-[100%] text-violet-400/90 z-30 relative" :icon="props.icon" />
+        <Icon class="size-[100%] text-violet-400/90 z-30 relative" :icon="icon" />
         <Icon
             class="size-[100%] text-violet-400/80 brightness-75 z-20 absolute top-[50%] translate-y-[-50%] blur-2xl"
-            :icon="props.icon"
+            :icon="icon"
         />
         <Icon
-            class="size-[100%] text-violet-400/60 brightness-75 z-20 absolute top-[51%] translate-y-[-50%] translate-x-[1.5%]"
-            :icon="props.icon"
+            class="size-[100%] text-violet-400/60 brightness-75 z-20 absolute top-[calc(50%+var(--yStep))] translate-y-[-50%] translate-x-[var(--xStep)]"
+            :icon="icon"
         />
         <Icon
-            class="size-[100%] text-violet-400/40 brightness-50 z-10 absolute top-[52%] translate-y-[-50%] translate-x-[3%]"
-            :icon="props.icon"
+            class="size-[100%] text-violet-400/40 brightness-50 z-10 absolute top-[calc(50%+var(--yStep)*2)] translate-y-[-50%] translate-x-[calc(var(--xStep)*2)]"
+            :icon="icon"
         />
     </div>
 </template>
@@ -19,9 +19,16 @@
 <script setup lang="ts">
 import { Icon } from "@iconify/vue";
 
-const props = defineProps<{
+const { icon, yStep = 2, xStep = 2 } = defineProps<{
     icon: string;
-    yStep: number;
-    xStep: number;
+    yStep?: number;
+    xStep?: number;
 }>();
 </script>
+
+<style scoped>
+div {
+    --xStep: calc(v-bind(xStep) * 1%);
+    --yStep: calc(v-bind(yStep) * 1%);
+}
+</style>
